bugfix: Avoid underflow as leakage trackers are decremeneted (#26)
The leakage trackers can go below 0 for two reasons: 1) numerical
precision issues when decrementing the trackers; 2) when the forward
pass that populates the trackers is not executed due to the absence of
heralding gates. In both cases it is safe to implement a hard limit on
the tracker underflowing below 0.